Initial commit
This is the initial implementation of Bailey, a simple, dynamically
typed language. This implementation uses a tree-walking interpreter
that reuses the host stack (i.e. the Rust stack) but keeps the heap in
a hash map (i.e. it does not directly use the Rust heap). Later, I
will add reference counting to the interpreter so memory can be
reclaimed.