Make interpreter use let-normalized form
This changes the interpreter so it uses the let-normalized form of the
program instead of the raw source form. This simplifies the handling
of reference counting, since all intermediate values are assigned to a
call stack variable. Moreover, the interpreter now has an explicit
call stack to hold the variables as opposed to partially using the
host stack for that. However, it does still use the host stack to keep
track of return locations.