Actually bump reference counts
This change adds code to actually bump the reference counts,
previously that code was missing. There are currently three cases when
the reference count gets bumped: (1) when initially allocating the
value and binding the first reference to it (2) when creating a tuple
or closure that holds a new reference to an existing value (3) when
mutating the fields of a tuple to point one of the tuple fields to a
(potentially different) existing value.