Use flat instruction space, add fibonacci test
The compiler now produces an intermediate representation that has a
flat representation for the instruction space (consisting of a list
blocks where each block has a sequence of instructions) and explicit
code pointers to refer to target locations.

This change also adds a first simple end-to-end test: a program for
computing Fibonacci numbers. It is able to correctly evaluate the nth
Fibonacci number now and leave the heap/stack empty at the end of the
program, so reference counting and basic stack/closure handling at
least seems to be working properly. Of course, needs more testing to
increase confidence and test more edge cases.