Add frame layout code
This adds code to determine the frame layout for variables that are
going to be allocated on the stack. This is an intermediate step in
targeting a language (either another intermediate language or
assembly) where variable accesses are performed by indexing into a
flat memory space as opposed to using host-language HashMaps and
string-based variable lookup.